Create PriceUpdater, move updatePrices() to PriceUpdater
